modules/client/requirements.txt
requests
numpy
opencv-python
pillow
pytest
pytest-mock
azure-iot-device
